requirement_video_clip.txt
absl-py==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
alembic==1.16.4
antlr4-python3-runtime==4.9.3
asteroid-filterbanks==0.4.0
async-timeout==5.0.1
attrs==25.3.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
colorlog==6.9.0
contourpy==1.3.2
cycler==0.12.1
decorator==4.4.2
docopt==0.6.2
einops==0.8.1
filelock==3.13.1
flatbuffers==25.2.10
fonttools==4.59.1
frozenlist==1.7.0
fsspec==2024.6.1
greenlet==3.2.4
huggingface-hub==0.34.4
HyperPyYAML==1.2.2
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
jax==0.6.2
jaxlib==0.6.2
Jinja2==3.1.4
joblib==1.5.1
julius==0.2.7
kiwisolver==1.4.9
lightning==2.5.3
lightning-utilities==0.15.2
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==2.1.5
matplotlib==3.10.5
mdurl==0.1.2
mediapipe==0.10.14
ml_dtypes==0.5.3
moviepy==1.0.3
mpmath==1.3.0
multidict==6.6.4
networkx==3.3
numpy==2.1.2
omegaconf==2.3.0
opencv-contrib-python==4.12.0.88
opencv-python==4.12.0.88
opt_einsum==3.4.0
optuna==4.5.0
packaging==25.0
pandas==2.3.1
pillow==11.0.0
primePy==1.3
proglog==0.1.12
propcache==0.3.2
protobuf==4.25.8
pyannote.audio==3.3.2
pyannote.core==5.0.0
pyannote.database==5.1.3
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
pycparser==2.22
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytorch-lightning==2.5.3
pytorch-metric-learning==2.9.0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.5
rich==14.1.0
ruamel.yaml==0.18.15
ruamel.yaml.clib==0.2.12
scikit-learn==1.7.1
scipy==1.15.3
semver==3.0.4
sentencepiece==0.2.1
shellingham==1.5.4
six==1.17.0
sortedcontainers==2.4.0
sounddevice==0.5.2
soundfile==0.13.1
speechbrain==1.0.3
SQLAlchemy==2.0.43
sympy==1.13.3
tabulate==0.9.0
tensorboardX==2.6.4
threadpoolctl==3.6.0
tomli==2.2.1
torch==2.8.0+cu128
torch-audiomentations==0.12.0
torch_pitch_shift==1.2.5
torchaudio==2.8.0
torchmetrics==1.8.1
torchvision==0.23.0+cu128
tqdm==4.67.1
typer==0.16.1
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.5.0
yarl==1.20.1
yt-dlp==2025.8.20
